+++ src/humanoid_entity.hpp
@@ -6,8 +6,8 @@ class MppHumanoidEntity : public MppMobEntity
 {
 public:
     MppHumanoidEntity();
-    float GetPhysicsWidth() const override;
-    float GetPhysicsHeight() const override;
+    int GetPhysicsWidth() const override;
+    int GetPhysicsHeight() const override;
     
 protected:
     int GetPhysicsOffsetX() const override;
