+++ src/slime_entity.hpp
@@ -8,8 +8,8 @@ class MppSlimeEntity : public MppMobEntity
 
 public:
     MppSlimeEntity();
-    float GetPhysicsWidth() const override;
-    float GetPhysicsHeight() const override;
+    int GetPhysicsWidth() const override;
+    int GetPhysicsHeight() const override;
     
 protected:
     int GetPhysicsOffsetX() const override;
